ANOVA for factorial designs

A wrapper around afex::aov_ez wich allows mix effects model.

Usage

aov_r(
  data,
  response,
  between = NULL,
  within = NULL,
  rowid = NULL,
  ss_type = 3,
  effsize.type = "unbiased",
  sphericity = "auto",
  conf.level = 0.95,
  character.only = FALSE,
  markdown
)

Arguments

data

Your dataset in long format, can have some missing values.

response

Response variable, numeric.

between

Character indicating the between-subject(s) factor(s)/column(s) in data. Default is NULL indicating no between-subjects factors. Must be character vector if more than one between-subject(s) factor(s)/column(s) is specified.

within

Character indicating the within-subject(s) factor(s)/column(s) in data. Default is NULL indicating no between-subjects factors. Must be character vector if more than one within-subject(s) factor(s)/column(s) is specified.

rowid

Character (of length 1) indicating the subject identifier column in data.

ss_type

Type of sum of squares for repeated measures ANOVA (defaults to 3). Possible values are "II", "III", 2, or 3.

effsize.type

The effect size used to estimate the effects of the factors on the response variable. Possible values are "eta" ("biased") or "omega" ("unbiased", the default).

sphericity

If "none", then sphericity assumption is assumed to be met for within-subject(s) factor(s). "GG": applies Greenhouse-Geisser correction. "HF": applies Hyunh-Feldt correction. 'auto' (Default) choose the appropiate correction based on Mauchly test of sphericity (p-value > 0.05)

conf.level

Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

character.only

Logical. checks whether to use the unevaluated expression or its content (when TRUE), asumming is a character vector. Defaults to FALSE.

markdown

Logical (default FALSE). If lbl is TRUE, then this argument specify if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).

Value

A list with statistical results.