Format submissions

Format submissions to various academic conferences and journals.

gen_paper(
  ...,
  engine = "xelatex",
  journal = "aom",
  keep_tex = TRUE,
  citation_package = "natbib",
  md_extensions = c("-autolink_bare_uris")
)

Arguments

...
Arguments passed to pdf_document().

engine
character. Latex engine to be used to produce the pdf.

journal
character. Journal to which the paper is submitted. See details for the available options.

keep_tex
Keep the intermediate tex file used in the conversion to PDF

citation_package
The LaTeX package to process citations, natbib or biblatex. Use default if neither package is to be used, which means citations will be processed via the command pandoc-citeproc.

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

Value

R Markdown output format to pass to render

Details

The templates currently available are:

aos for Accounting, Organizations, and Society

bria for Behavioral Research in Accounting

car for Contemporary Accounting Research

cpa for Critical Perspectives on Accounting

ear for European Accounting Review

jae for Journal of Accounting and Economics

jar for Journal of Accounting Research

jaaf for Journal of Accounting, Auditing, and Finance

jmar for Journal of Management Accounting Research

mar for Management Accounting Research

tar for The Accounting Review

jfe for Journal of Financial Economics

amj for Academy of Management Journal

amr for Academy of Management Review

jibs for Journal of International Business Studies

jom for Journal of Management

mnsc for Management Science

orsc for Organization Science

smj for Strategic Management Journal

orm for Organizational Research Methods

jap for Journal of Applied Psychology

Author

Nicolas Mangin