Generate report-ready labels for statistical report

A list containing stats, p value, effectsize, confidence/credible interval and a concatenated string named 'full'.

Usage

lablr(t, markdown = FALSE)

Arguments

t

Output from any of the functions, k_sample, two_sample or one_sample.

markdown

Logical (default FALSE), indicating if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).