Two Sample test

A list containing results from a two-sample test and effect size plus confidence intervals.

two_sample(
  data,
  x,
  y,
  rowid = NULL,
  type = "auto",
  paired = FALSE,
  var.equal = FALSE,
  trim = 0.2,
  nboot = 100L,
  effsize.type = "unbiased",
  alternative = "two.sided",
  conf.level = 0.95,
  lbl = if (is.null(markdown)) FALSE else TRUE,
  markdown = NULL,
  internal = FALSE,
  ...
)

Arguments

data
Data frame from which x and y (and possibly rowid if provided) will be searched.

x
Character name for the grouping factor. Must be present in data

y
Character name for the response variable. Must be present in data.

rowid
Character name for the subject-id column. If null, then is assumed that data is sorted for paired designs, creating one. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate when there are more than two levels in x and there are NAs present.

type
Set "auto" (default) for checking the normality and homogeneity of variances for test selection. Other options are "p" for parametric, "np" for non-parametric and "r" for robust tests.

paired
Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is FALSE.

var.equal
Logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

trim
Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

nboot
Number of bootstrap samples for computing confidence interval for the effect size (Default: 100L).

effsize.type
Options are "unbiased" or "g" for Hedges g and "biased" or "d" for Cohen's d as a measure of effect size. rank-biserial correlation is used for non-parametric analysis.

alternative
A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".

conf.level
Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

lbl
Logical (default FALSE) indicating if a report ready output is desired. This will change the output to a list with characters rather than numeric vectors.

markdown
Logical (default FALSE). If lbl is TRUE, then this argument specify if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).

internal
Logical to whether this function is being used inside of internal functions.

...
Currently ignored.