One Sample test

A list containing results from a one-sample test.

Usage

one_sample(
  data,
  y,
  type,
  test.value = 0,
  effsize.type = "unbiased",
  alternative = "two.sided",
  conf.level = 0.95,
  character.only = FALSE,
  markdown,
  ...
)

Arguments

data

data frame from which x and y (and possibly rowid if provided) will be searched.

y

Character for the response variable Must be present in data

type

set "auto" (default) for checking the normality for test selection. Other options are "p" for parametric, "np" for non-parametric and "r" for robust tests.

test.value

A number indicating the true value of the mean (Default: 0) to be tested.

effsize.type

options are "unbiased" or "g" for Hedges g and "biased" or "d" for Cohen's d as a measure of effect size. rank-biserial correlation is used for non-parametric analysis.

alternative

A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".

conf.level

Confidence/Credible Interval (CI) level. Default to 0.95 (95%).

character.only

whether to treat x as a character. Default is FALSE.

markdown

Logical (default FALSE). If lbl is TRUE, then this argument specify if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).

...

Currently ignored.